# HotelBookingDetails/HotelBookingDetails.WasmClient/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "HotelBookingDetails.WebApi",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Room/{id}": {
      "get": {
        "tags": [
          "Room"
        ],
        "summary": "Запрос комнаты по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Room"
        ],
        "summary": "Запрос на изменение комнаты по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Room"
        ],
        "summary": "Удаление комнаты по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TypeRoom": {
      "get": {
        "tags": [
          "TypeRoom"
        ],
        "summary": "Запрос возвращающий список типов комнат",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TypeRoom"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TypeRoom"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TypeRoom"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TypeRoom"
        ],
        "summary": "Запрос на добавления типа комнаты",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TypeRoomDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TypeRoomDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TypeRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TypeRoom/{id}": {
      "get": {
        "tags": [
          "TypeRoom"
        ],
        "summary": "Запрос типа комнаты по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TypeRoom"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeRoom"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeRoom"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TypeRoom"
        ],
        "summary": "Запрос на изменение типа комнаты по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TypeRoomDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TypeRoomDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TypeRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "TypeRoom"
        ],
        "summary": "Удаление типа комнаты по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Hotel": {
        "required": [
          "address",
          "city",
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "address": {
            "minLength": 1,
            "type": "string"
          },
          "city": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Room": {
        "required": [
          "capacity",
          "cost",
          "hotel",
          "id",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/TypeRoom"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "cost": {
            "type": "integer",
            "format": "int32"
          },
          "hotel": {
            "$ref": "#/components/schemas/Hotel"
          }
        },
        "additionalProperties": false
      },
      "RoomDto": {
        "type": "object",
        "properties": {
          "typeId": {
            "type": "integer",
            "format": "int32"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "cost": {
            "type": "integer",
            "format": "int32"
          },
          "hotelId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "TypeRoom": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "TypeRoomDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}